Rename start to run
`runc run` is the command that will create and start a container in one
single command.
