+++ tests/integration/cgroups.bats
@@ -47,8 +47,8 @@ EOF
     DATA=$(echo ${DATA} | sed 's/\n/\\n/g')
     sed -i "s/\(\"resources\": {\)/\1\n${DATA}/" ${BUSYBOX_BUNDLE}/config.json
 
-    # start a detached busybox to work with
-    runc start -d --console /dev/pts/ptmx test_cgroups_kmem
+    # run a detached busybox to work with
+    runc run -d --console /dev/pts/ptmx test_cgroups_kmem
     [ "$status" -eq 0 ]
     wait_for_container 15 1 test_cgroups_kmem
 
@@ -64,8 +64,8 @@ EOF
     # Add cgroup path
     sed -i 's/\("linux": {\)/\1\n    "cgroupsPath": "runc-cgroups-integration-test",/'  ${BUSYBOX_BUNDLE}/config.json
 
-    # start a detached busybox to work with
-    runc start -d --console /dev/pts/ptmx test_cgroups_kmem
+    # run a detached busybox to work with
+    runc run -d --console /dev/pts/ptmx test_cgroups_kmem
     [ "$status" -eq 0 ]
     wait_for_container 15 1 test_cgroups_kmem
