+++ tests/integration/debug.bats
@@ -12,15 +12,15 @@ function teardown() {
 }
 
 @test "global --debug" {
-  # start hello-world
-  runc --debug start test_hello
+  # run hello-world
+  runc --debug run test_hello
   echo "${output}"
   [ "$status" -eq 0 ]
 }
 
 @test "global --debug to --log" {
-  # start hello-world
-  runc --log log.out --debug start test_hello
+  # run hello-world
+  runc --log log.out --debug run test_hello
   [ "$status" -eq 0 ]
 
   # check output does not include debug info
@@ -36,8 +36,8 @@ function teardown() {
 }
 
 @test "global --debug to --log --log-format 'text'" {
-  # start hello-world
-  runc --log log.out --log-format "text" --debug start test_hello
+  # run hello-world
+  runc --log log.out --log-format "text" --debug run test_hello
   [ "$status" -eq 0 ]
 
   # check output does not include debug info
@@ -53,8 +53,8 @@ function teardown() {
 }
 
 @test "global --debug to --log --log-format 'json'" {
-  # start hello-world
-  runc --log log.out --log-format "json" --debug start test_hello
+  # run hello-world
+  runc --log log.out --log-format "json" --debug run test_hello
   [ "$status" -eq 0 ]
 
   # check output does not include debug info
