+++ tests/integration/kill.bats
@@ -14,8 +14,8 @@ function teardown() {
 
 @test "kill detached busybox" {
 
-  # start busybox detached
-  runc start -d --console /dev/pts/ptmx test_busybox
+  # run busybox detached
+  runc run -d --console /dev/pts/ptmx test_busybox
   [ "$status" -eq 0 ]
 
   # check state
