+++ tests/integration/list.bats
@@ -18,16 +18,16 @@ function teardown() {
 }
 
 @test "list" {
-  # start a few busyboxes detached
-  ROOT=$HELLO_BUNDLE runc start -d --console /dev/pts/ptmx test_box1
+  # run a few busyboxes detached
+  ROOT=$HELLO_BUNDLE runc run -d --console /dev/pts/ptmx test_box1
   [ "$status" -eq 0 ]
   wait_for_container_inroot 15 1 test_box1 $HELLO_BUNDLE
 
-  ROOT=$HELLO_BUNDLE runc start -d --console /dev/pts/ptmx test_box2
+  ROOT=$HELLO_BUNDLE runc run -d --console /dev/pts/ptmx test_box2
   [ "$status" -eq 0 ]
   wait_for_container_inroot 15 1 test_box2 $HELLO_BUNDLE
 
-  ROOT=$HELLO_BUNDLE runc start -d --console /dev/pts/ptmx test_box3
+  ROOT=$HELLO_BUNDLE runc run -d --console /dev/pts/ptmx test_box3
   [ "$status" -eq 0 ]
   wait_for_container_inroot 15 1 test_box3 $HELLO_BUNDLE
